Refactored MyApp
