+++ packages/react-frontend/src/MyApp.jsx
@@ -1,337 +1,32 @@
 import React, { useState, useEffect, useRef } from "react";
 import { useNavigate } from "react-router-dom";
-import PropTypes from "prop-types";
-
-// Simple Cute Bread Component (Stationary)
-const CuteBread = ({ size = "40px", className = "" }) => {
-  return (
-    <div className={`cute-bread ${className}`} style={{ width: size, height: size }}>
-      <div className="bread-body">
-        <div className="bread-eyes">
-          <div className="eye left-eye"></div>
-          <div className="eye right-eye"></div>
-        </div>
-        <div className="bread-mouth"></div>
-        <div className="bread-cheeks">
-          <div className="cheek left-cheek"></div>
-          <div className="cheek right-cheek"></div>
-        </div>
-      </div>
-    </div>
-  );
-};
-
-CuteBread.propTypes = {
-  size: PropTypes.string,
-  className: PropTypes.string,
-};
-
-// Calendar Component
-const CalendarView = ({ tasks, showButterTasks }) => {
-  const [currentMonth, setCurrentMonth] = useState(new Date().getMonth());
-  const [currentYear, setCurrentYear] = useState(new Date().getFullYear());
-  const [selectedDate, setSelectedDate] = useState(null);
-
-  const months = [
-    "January", "February", "March", "April", "May", "June",
-    "July", "August", "September", "October", "November", "December"
-  ];
-
-  // Get urgency color
-  const getUrgencyColor = (urgency) => {
-    if (urgency <= 2) return "#28a745"; // Green (low)
-    if (urgency <= 4) return "#ffc107"; // Yellow
-    if (urgency <= 6) return "#fd7e14"; // Orange
-    if (urgency <= 8) return "#dc3545"; // Red
-    return "#6f42c1"; // Purple (highest)
-  };
-
-  // Get tasks for a specific date
-  const getTasksForDate = (date) => {
-    const dateStr = date.toISOString().split('T')[0];
-    let filteredTasks = tasks.filter(task => {
-      if (!task.due_date) return false;
-      const taskDate = new Date(task.due_date).toISOString().split('T')[0];
-      return taskDate === dateStr;
-    });
-
-    // Filter out butter tasks if setting is disabled
-    if (!showButterTasks) {
-      filteredTasks = filteredTasks.filter(task => Number(task.ease) >= 60);
-    }
-
-    return filteredTasks;
-  };
-
-  // Generate calendar days
-  const getDaysInMonth = () => {
-    const firstDay = new Date(currentYear, currentMonth, 1);
-    const lastDay = new Date(currentYear, currentMonth + 1, 0);
-    const startDate = new Date(firstDay);
-    startDate.setDate(startDate.getDate() - firstDay.getDay());
-
-    const days = [];
-    const endDate = new Date(startDate);
-    endDate.setDate(endDate.getDate() + 41); // 6 weeks
-
-    for (let date = new Date(startDate); date < endDate; date.setDate(date.getDate() + 1)) {
-      days.push(new Date(date));
-    }
-    return days;
-  };
-
-  const days = getDaysInMonth();
-  const today = new Date();
-
-  const navigateMonth = (direction) => {
-    if (direction === 'prev') {
-      if (currentMonth === 0) {
-        setCurrentMonth(11);
-        setCurrentYear(currentYear - 1);
-      } else {
-        setCurrentMonth(currentMonth - 1);
-      }
-    } else {
-      if (currentMonth === 11) {
-        setCurrentMonth(0);
-        setCurrentYear(currentYear + 1);
-      } else {
-        setCurrentMonth(currentMonth + 1);
-      }
-    }
-  };
-
-  return (
-    <div className="calendar-container">
-      <div className="calendar-header">
-        <h2>📆 Calendar View</h2>
-        <div className="calendar-navigation">
-          <button className="nav-btn" onClick={() => navigateMonth('prev')}>&lt;</button>
-          <span className="calendar-month-year">
-            {months[currentMonth]} {currentYear}
-          </span>
-          <button className="nav-btn" onClick={() => navigateMonth('next')}>&gt;</button>
-        </div>
-      </div>
-
-      <div className="calendar-grid">
-        <div className="calendar-weekdays">
-          {['Sun', 'Mon', 'Tue', 'Wed', 'Thu', 'Fri', 'Sat'].map(day => (
-            <div key={day} className="calendar-weekday">{day}</div>
-          ))}
-        </div>
-
-        <div className="calendar-days">
-          {days.map((date, index) => {
-            const dayTasks = getTasksForDate(date);
-            const isCurrentMonth = date.getMonth() === currentMonth;
-            const isToday = date.toDateString() === today.toDateString();
-            const isSelected = selectedDate && date.toDateString() === selectedDate.toDateString();
-
-            return (
-              <div
-                key={index}
-                className={`calendar-day ${!isCurrentMonth ? 'other-month' : ''} ${isToday ? 'today' : ''} ${isSelected ? 'selected' : ''}`}
-                onClick={() => setSelectedDate(date)}
-              >
-                <div className="calendar-day-number">{date.getDate()}</div>
-                {dayTasks.length > 0 && (
-                  <div className="calendar-tasks">
-                    {dayTasks.slice(0, 3).map((task, i) => (
-                      <div
-                        key={i}
-                        className="calendar-task-dot"
-                        style={{ backgroundColor: getUrgencyColor(task.urgency) }}
-                        title={`${task.name} (Urgency: ${task.urgency})`}
-                      />
-                    ))}
-                    {dayTasks.length > 3 && (
-                      <div className="calendar-task-more">+{dayTasks.length - 3}</div>
-                    )}
-                  </div>
-                )}
-              </div>
-            );
-          })}
-        </div>
-      </div>
-
-      {/* Selected Date Details */}
-      {selectedDate && (
-        <div className="calendar-selected-date">
-          <h3>Tasks for {selectedDate.toLocaleDateString()}</h3>
-          {getTasksForDate(selectedDate).length === 0 ? (
-            <p>No tasks scheduled for this date.</p>
-          ) : (
-            <div className="calendar-task-list">
-              {getTasksForDate(selectedDate).map(task => (
-                <div key={task._id} className="calendar-task-item">
-                  <div
-                    className="task-urgency-dot"
-                    style={{ backgroundColor: getUrgencyColor(task.urgency) }}
-                  />
-                  <div className="task-details">
-                    <strong>{task.name}</strong>
-                    <div className="task-meta">
-                      {task.category} • {task.ease} min • Urgency: {task.urgency}/10
-                    </div>
-                    {task.description && <p className="task-description">{task.description}</p>}
-                  </div>
-                </div>
-              ))}
-            </div>
-          )}
-        </div>
-      )}
-
-      {/* Legend */}
-      <div className="calendar-legend">
-        <h4>Urgency Scale:</h4>
-        <div className="legend-items">
-          <div className="legend-item">
-            <div className="legend-dot" style={{ backgroundColor: "#28a745" }}></div>
-            <span>Low (1-2)</span>
-          </div>
-          <div className="legend-item">
-            <div className="legend-dot" style={{ backgroundColor: "#ffc107" }}></div>
-            <span>Medium (3-4)</span>
-          </div>
-          <div className="legend-item">
-            <div className="legend-dot" style={{ backgroundColor: "#fd7e14" }}></div>
-            <span>High (5-6)</span>
-          </div>
-          <div className="legend-item">
-            <div className="legend-dot" style={{ backgroundColor: "#dc3545" }}></div>
-            <span>Urgent (7-8)</span>
-          </div>
-          <div className="legend-item">
-            <div className="legend-dot" style={{ backgroundColor: "#6f42c1" }}></div>
-            <span>Critical (9-10)</span>
-          </div>
-        </div>
-      </div>
-    </div>
-  );
-};
-
-CalendarView.propTypes = {
-  tasks: PropTypes.array.isRequired,
-  showButterTasks: PropTypes.bool,
-};
-
-const EditableCategory = ({ name, index, onNameChange, taskCount, onClick }) => {
-  const [editing, setEditing] = useState(false);
-  const [tempName, setTempName] = useState(name);
-
-  const handleBlur = () => {
-    setEditing(false);
-    onNameChange(index, tempName);
-  };
-
-  const handleKeyDown = (e) => {
-    if (e.key === "Enter") {
-      e.preventDefault();
-      setEditing(false);
-      onNameChange(index, tempName);
-    }
-  };
-
-  return editing ? (
-    <input
-      type="text"
-      value={tempName}
-      autoFocus
-      onChange={(e) => setTempName(e.target.value)}
-      onBlur={handleBlur}
-      onKeyDown={handleKeyDown}
-      className="category-input"
-    />
-  ) : (
-    <div className="category-box" style={{ cursor: "pointer" }}>
-      <h3 
-        onClick={(e) => {
-          e.stopPropagation();
-          setEditing(true);
-        }}
-        style={{ cursor: "text" }}
-      >
-        {name || <em>Click to name me</em>}
-      </h3>
-      {name && (
-        <p 
-          className="task-count"
-          onClick={(e) => {
-            e.stopPropagation();
-            onClick();
-          }}
-          style={{ cursor: "pointer" }}
-        >
-          ({taskCount} tasks)
-        </p>
-      )}
-    </div>
-  );
-};
+import CuteBread from "./CuteBread";
+import CalendarView from "./CalendarView";
+import EditableCategory from "./EditableCategory";
+import { useSettings } from "./useSettings";
+import { useTasks } from "./useTasks";
+import { useProfile } from "./useProfile";
+import { getAchievements, breadQuotes } from "./achievements";
 
 function MyApp() {
-  const [tasks, setTasks] = useState([]);
-  const [checkedTasks, setCheckedTasks] = useState([]);
-  const [progress, setProgress] = useState(0);
   const [selectedCategory, setSelectedCategory] = useState(null);
   const [sidebarOpen, setSidebarOpen] = useState(false);
   const [activePage, setActivePage] = useState("home");
   const navigate = useNavigate();
   const sidebarRef = useRef();
 
-  // Settings state
-  const [settings, setSettings] = useState(() => {
-    const savedSettings = localStorage.getItem('crumblist-settings');
-    return savedSettings ? JSON.parse(savedSettings) : {
-      darkMode: false,
-      butterThreshold: 60,
-      showButterInCalendar: true,
-      dailyQuotes: false,
-      colorTheme: 'pink'
-    };
-  });
+  // Custom hooks
+  const { settings, updateSetting, requestNotificationPermission, clearCompletedTasks, resetCategoriesToDefault } = useSettings();
+  const { tasks, checkedTasks, progress, toggleChecked, completeCheckedTasks } = useTasks();
+  const { userProfile, editingProfile, updateProfile, handleProfileKeyDown } = useProfile();
 
-  // User profile state
-  const [userProfile, setUserProfile] = useState(() => {
-    const savedProfile = localStorage.getItem('crumblist-profile');
-    return savedProfile ? JSON.parse(savedProfile) : {
-      name: 'Your Name',
-      title: 'Task Master'
-    };
-  });
-
-  const [editingProfile, setEditingProfile] = useState({ name: false, title: false });
-
-  // Bread-themed quotes
-  const breadQuotes = [
-    "Stay focused and keep toasting.",
-    "Small crumbs lead to big loaves.",
-    "No task is too crusty to conquer.",
-    "Rise to the occasion like good dough.",
-    "Every great achievement starts with a single grain.",
-    "Knead your goals until they're perfect.",
-    "Life is what you bake it.",
-    "Don't loaf around - get things done!",
-    "You're the yeast that makes things rise.",
-    "Success is a recipe worth following.",
-    "Proof that hard work pays off.",
-    "Butter believe in yourself!",
-    "Time to roll up your sleeves and get baking.",
-    "Fresh starts are like warm bread - always better.",
-    "Slice through challenges one task at a time."
-  ];
-
-  // Random quote that changes on each render/refresh
+  // Quote state
   const [quote] = useState(() => {
     const randomQuote = breadQuotes[Math.floor(Math.random() * breadQuotes.length)];
     return `${randomQuote} – Mr. Crumb`;
   });
 
-  // Initialize categories from localStorage or use defaults
+  // Categories state
   const [categories, setCategories] = useState(() => {
     const savedCategories = localStorage.getItem('crumblist-categories');
     if (savedCategories) {
@@ -347,6 +42,7 @@ function MyApp() {
     ];
   });
 
+  // Task filtering
   const butterTasks = tasks.filter((task) => Number(task.ease) < settings.butterThreshold);
   const normalTasks = tasks.filter((task) => Number(task.ease) >= settings.butterThreshold);
 
@@ -358,194 +54,36 @@ function MyApp() {
     );
   });
 
-  // Apply theme class to body
+  // Save categories
+  useEffect(() => {
+    localStorage.setItem('crumblist-categories', JSON.stringify(categories));
+  }, [categories]);
+
+  // Apply themes
   useEffect(() => {
     document.body.className = `theme-${settings.colorTheme} ${settings.darkMode ? 'dark-mode' : ''}`;
-    
-    // Apply theme to pink-background class
     const pinkBg = document.querySelector('.pink-background');
     if (pinkBg) {
       pinkBg.className = `pink-background theme-${settings.colorTheme} ${settings.darkMode ? 'dark-mode' : ''}`;
     }
   }, [settings.darkMode, settings.colorTheme]);
 
-  // Save settings to localStorage
-  useEffect(() => {
-    localStorage.setItem('crumblist-settings', JSON.stringify(settings));
-  }, [settings]);
-
-  // Save profile to localStorage
-  useEffect(() => {
-    localStorage.setItem('crumblist-profile', JSON.stringify(userProfile));
-  }, [userProfile]);
-
-  // Daily quote notification
-  useEffect(() => {
-    if (settings.dailyQuotes) {
-      const lastQuoteDate = localStorage.getItem('last-quote-date');
-      const today = new Date().toDateString();
-      
-      if (lastQuoteDate !== today) {
-        const randomQuote = breadQuotes[Math.floor(Math.random() * breadQuotes.length)];
-        if (Notification.permission === 'granted') {
-          new Notification('Daily Motivation from Mr. Crumb', {
-            body: randomQuote,
-            icon: '🍞'
-          });
-        }
-        localStorage.setItem('last-quote-date', today);
-      }
-    }
-  }, [settings.dailyQuotes]);
-
-  // Request notification permission
-  const requestNotificationPermission = () => {
-    if ('Notification' in window) {
-      Notification.requestPermission();
-    }
-  };
-
-  // Settings update functions
-  const updateSetting = (key, value) => {
-    setSettings(prev => ({ ...prev, [key]: value }));
-  };
-
-  const clearCompletedTasks = () => {
-    if (window.confirm('Are you sure you want to clear all completed tasks? This cannot be undone.')) {
-      console.log('Clearing completed tasks...');
-      alert('Completed tasks cleared!');
-    }
-  };
-
-  const updateProfile = (field, value) => {
-    setUserProfile(prev => ({ ...prev, [field]: value }));
-    setEditingProfile(prev => ({ ...prev, [field]: false }));
-  };
-
-  const handleProfileKeyDown = (e, field) => {
-    if (e.key === 'Enter') {
-      updateProfile(field, e.target.value);
-    }
-  };
-
-  // Calculate achievements
-  const getAchievements = () => {
-    const achievements = [];
-    
-    if (tasks.length >= 10) {
-      achievements.push({ icon: '🥖', text: 'Bread Collector', desc: 'Created 10+ tasks' });
-    }
-    
-    if (butterTasks.length >= 5) {
-      achievements.push({ icon: '🧈', text: 'Butter Master', desc: '5+ butter tasks' });
-    }
-    
-    if (progress === 100 && tasks.length > 0) {
-      achievements.push({ icon: '🔥', text: 'Perfectionist', desc: '100% completion' });
-    }
-    
-    if (checkedTasks.length >= 5) {
-      achievements.push({ icon: '⚡', text: 'Speed Demon', desc: '5+ tasks completed today' });
-    }
-
-    if (tasks.length >= 50) {
-      achievements.push({ icon: '🏆', text: 'Task Champion', desc: '50+ total tasks' });
-    }
-
-    if (categories.length > 6) {
-      achievements.push({ icon: '📚', text: 'Organizer', desc: 'Created custom categories' });
-    }
-
-    return achievements;
-  };
-
-  const resetCategoriesToDefault = () => {
-    if (window.confirm('Reset categories to default? This will remove any custom categories you\'ve created.')) {
-      const defaultCategories = [
-        { name: "School" },
-        { name: "Work" },
-        { name: "Errands" },
-        { name: "Health" },
-        { name: "Fitness" },
-        { name: "Chores" }
-      ];
-      setCategories(defaultCategories);
-      localStorage.setItem('crumblist-categories', JSON.stringify(defaultCategories));
-      alert('Categories reset to default!');
-    }
-  };
-
-  // Close sidebar when clicking outside
+  // Sidebar outside click
   useEffect(() => {
     function handleClickOutside(event) {
       if (sidebarRef.current && !sidebarRef.current.contains(event.target)) {
         setSidebarOpen(false);
       }
     }
-
     if (sidebarOpen) {
       document.addEventListener("mousedown", handleClickOutside);
     }
-
     return () => {
       document.removeEventListener("mousedown", handleClickOutside);
     };
   }, [sidebarOpen]);
 
-  // Fetch tasks from backend
-  const fetchTasks = () => {
-    fetch("https://crumblist-g5htfcg7afh8ehdw.canadacentral-01.azurewebsites.net/tasks")
-      .then((res) => res.json())
-      .then((json) => {
-        console.log("Fetched tasks:", json.task_list);
-        setTasks(json.task_list || []);
-      })
-      .catch((error) => {
-        console.error("Error fetching tasks:", error);
-        setTasks([]);
-      });
-  };
-
-  useEffect(() => {
-    fetchTasks();
-  }, []);
-
-  // Save categories to localStorage whenever they change
-  useEffect(() => {
-    localStorage.setItem('crumblist-categories', JSON.stringify(categories));
-  }, [categories]);
-
-  useEffect(() => {
-    const total = tasks.length;
-    const completed = checkedTasks.length;
-    const percent = total === 0 ? 100 : Math.round((completed / total) * 100);
-    setProgress(percent);
-  }, [checkedTasks, tasks]);
-
-  const toggleChecked = (taskId) => {
-    setCheckedTasks((prev) =>
-      prev.includes(taskId)
-        ? prev.filter((id) => id !== taskId)
-        : [...prev, taskId]
-    );
-  };
-
-  const completeCheckedTasks = () => {
-    const updatedTasks = tasks.filter((task) => {
-      return !checkedTasks.includes(task._id);
-    });
-
-    setTasks(updatedTasks);
-    setCheckedTasks([]);
-
-    // Delete from backend
-    checkedTasks.forEach((id) => {
-      fetch(`https://crumblist-g5htfcg7afh8ehdw.canadacentral-01.azurewebsites.net/tasks/${id}`, {
-        method: "DELETE",
-      }).catch((err) => console.error("Delete failed", err));
-    });
-  };
-
+  // Helper functions
   const updateCategoryName = (index, newName) => {
     const updated = [...categories];
     updated[index].name = newName;
@@ -563,10 +101,7 @@ function MyApp() {
   return (
     <div className={`pink-background theme-${settings.colorTheme} ${settings.darkMode ? 'dark-mode' : ''}`}>
       {/* Hamburger Menu */}
-      <div
-        className="hamburger-menu"
-        onClick={() => setSidebarOpen(!sidebarOpen)}
-      >
+      <div className="hamburger-menu" onClick={() => setSidebarOpen(!sidebarOpen)}>
         ☰
       </div>
 
@@ -717,7 +252,7 @@ function MyApp() {
                 </button>
               </div>
               <div className="setting-item">
-                <button className="settings-button warning-btn" onClick={resetCategoriesToDefault}>
+                <button className="settings-button warning-btn" onClick={() => resetCategoriesToDefault(setCategories)}>
                   🔄 Reset Categories to Default
                 </button>
               </div>
@@ -760,7 +295,7 @@ function MyApp() {
               ) : (
                 <h1 
                   className="profile-title"
-                  onClick={() => setEditingProfile(prev => ({ ...prev, title: true }))}
+                  onClick={() => updateProfile('editTitle', true)}
                 >
                   {userProfile.title}
                 </h1>
@@ -778,7 +313,7 @@ function MyApp() {
               ) : (
                 <p 
                   className="profile-name"
-                  onClick={() => setEditingProfile(prev => ({ ...prev, name: true }))}
+                  onClick={() => updateProfile('editName', true)}
                 >
                   {userProfile.name}
                 </p>
@@ -814,7 +349,7 @@ function MyApp() {
           <div className="profile-achievements">
             <h3>🏆 Achievements</h3>
             <div className="achievement-grid">
-              {getAchievements().map((achievement, index) => (
+              {getAchievements(tasks, butterTasks, progress, checkedTasks, categories).map((achievement, index) => (
                 <div key={index} className="achievement-badge">
                   <div className="badge-icon">{achievement.icon}</div>
                   <div className="badge-content">
@@ -823,7 +358,7 @@ function MyApp() {
                   </div>
                 </div>
               ))}
-              {getAchievements().length === 0 && (
+              {getAchievements(tasks, butterTasks, progress, checkedTasks, categories).length === 0 && (
                 <div className="no-achievements">
                   <p>Complete tasks to unlock achievements! 🎯</p>
                 </div>
@@ -962,13 +497,5 @@ function MyApp() {
     </div>
   );
 }
-
-EditableCategory.propTypes = {
-  name: PropTypes.string.isRequired,
-  index: PropTypes.number.isRequired,
-  onNameChange: PropTypes.func.isRequired,
-  taskCount: PropTypes.number,
-  onClick: PropTypes.func,
-};
 
 export default MyApp;
